changed timestamp filter props to uint64

+++ golang-backend/internal/handlers/handlers.go
@@ -35,6 +35,6 @@ func ToHandlerJSONResponse(w http.ResponseWriter, data any, err error, statusCod
 }
 
 type TimestampFilter struct {
-	From *int `json:"from"`
-	To   *int `json:"to"`
+	From *uint64 `json:"from"`
+	To   *uint64 `json:"to"`
 }
